# backend/swagger/swagger.yaml
openapi: 3.1.0
info:
  title: ASAP HVAC Sample API
  version: 0.2.0
  description: |
    Sample REST API for the ASAP HVAC web demo. Provides CRUD support for
    HVAC operational data with pagination, filtering, and validation via Zod
    schemas. Designed to be upgraded to relational storage in future iterations.
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /api/service-offerings:
    get:
      summary: List service offerings
      tags: [Service offerings]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: category
          in: query
          description: Filter by service category.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: name
          in: query
          description: Filter by service name (case-insensitive match).
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of service offerings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedServiceOffering'
    post:
      summary: Create a new service offering
      tags: [Service offerings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceOfferingInput'
      responses:
        '201':
          description: Newly created service offering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOffering'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/service-offerings/{id}:
    parameters:
      - $ref: '#/components/parameters/ServiceOfferingId'
    get:
      summary: Retrieve a service offering
      tags: [Service offerings]
      responses:
        '200':
          description: Service offering details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOffering'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update service offering
      tags: [Service offerings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceOfferingUpdate'
      responses:
        '200':
          description: Updated service offering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOffering'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a service offering
      tags: [Service offerings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceOfferingUpdate'
      responses:
        '200':
          description: Updated service offering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOffering'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a service offering
      tags: [Service offerings]
      responses:
        '204':
          description: Service offering deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/quote-requests:
    get:
      summary: List quote requests
      tags: [Quote requests]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          description: Filter by quote request status.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/QuoteStatus'
          style: form
          explode: true
        - name: urgency
          in: query
          description: Filter by urgency level.
          schema:
            type: array
            items:
              type: string
              enum: [low, normal, high]
          style: form
          explode: true
        - name: serviceType
          in: query
          description: Filter by requested service type.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: customerName
          in: query
          description: Filter by customer name.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of quote requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuoteRequest'
    post:
      summary: Create a quote request
      tags: [Quote requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequestInput'
      responses:
        '201':
          description: Newly created quote request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteRequest'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/quote-requests/{id}:
    parameters:
      - $ref: '#/components/parameters/QuoteRequestId'
    get:
      summary: Retrieve a quote request
      tags: [Quote requests]
      responses:
        '200':
          description: Quote request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update quote request
      tags: [Quote requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequestUpdate'
      responses:
        '200':
          description: Updated quote request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteRequest'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a quote request
      tags: [Quote requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequestUpdate'
      responses:
        '200':
          description: Updated quote request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteRequest'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a quote request
      tags: [Quote requests]
      responses:
        '204':
          description: Quote request deleted successfully
        '400':
          description: Quote request cannot be deleted when dependent records exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based indexing).
      schema:
        type: integer
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Number of records per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: Sort expression in the format `<field>[:asc|desc]`.
      schema:
        type: string
...
    ServiceOfferingId:
      name: id
      in: path
      required: true
      description: Unique identifier for the service offering
      schema:
        type: string
...
    QuoteRequestId:
      name: id
      in: path
      required: true
      description: Unique identifier for the quote request
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Customer email must be unique.
        details:
          description: Optional error metadata
          nullable: true
    ListMeta:
      type: object
      required: [total, page, limit, hasNextPage]
      properties:
        total:
          type: integer
          example: 45
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 25
        hasNextPage:
          type: boolean
          example: true
...
    PaginatedServiceOffering:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOffering'
        meta:
          $ref: '#/components/schemas/ListMeta'
    ServiceOffering:
      type: object
      required: [id, name, category, basePriceRange, description]
      properties:
        id:
          type: string
          example: svc-diagnostic
        name:
          type: string
        category:
          type: string
        basePriceRange:
          type: string
        description:
          type: string
    ServiceOfferingInput:
      type: object
      required: [name, category, basePriceRange, description]
      properties:
        id:
          type: string
          description: Optional custom identifier. If omitted, an id is generated.
        name:
          type: string
        category:
          type: string
        basePriceRange:
          type: string
        description:
          type: string
    ServiceOfferingUpdate:
      allOf:
        - $ref: '#/components/schemas/ServiceOfferingInput'
...
    QuoteStatus:
      type: string
      enum: [new, awaiting-scheduling, scheduled, inspection-complete, awaiting-approval, approved, declined]
    PaginatedQuoteRequest:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuoteRequest'
        meta:
          $ref: '#/components/schemas/ListMeta'
    QuoteRequest:
      type: object
      required:
        - id
        - customerName
        - contactPhone
        - email
        - address
        - serviceType
        - urgency
        - requestedDate
        - symptoms
        - status
      properties:
        id:
          type: string
        customerName:
          type: string
        contactPhone:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        serviceType:
          type: string
        urgency:
          type: string
          enum: [low, normal, high]
        requestedDate:
          type: string
          format: date
        unitAgeYears:
          type: number
          nullable: true
        symptoms:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/QuoteStatus'
    QuoteRequestInput:
      type: object
      required:
        - customerName
        - contactPhone
        - email
        - address
        - serviceType
        - requestedDate
      properties:
        id:
          type: string
        customerName:
          type: string
        contactPhone:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        serviceType:
          type: string
        urgency:
          type: string
          enum: [low, normal, high]
        requestedDate:
          type: string
          format: date
        unitAgeYears:
          type: number
          nullable: true
        symptoms:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/QuoteStatus'
    QuoteRequestUpdate:
      allOf:
        - $ref: '#/components/schemas/QuoteRequestInput'
...